# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: ParsePilot AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Job application management
  - name: analysis
    description: CV analysis and AI operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to returnTo on success.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all job applications for the current user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    put:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /applications/{id}/tailored-cv:
    patch:
      operationId: saveTailoredCv
      tags: [applications]
      summary: Save edited tailored CV text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTailoredCvBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found

  /applications/{id}/cover-letter-save:
    patch:
      operationId: saveCoverLetter
      tags: [applications]
      summary: Save edited cover letter text
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCoverLetterBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found

  /applications/{id}/analyze:
    post:
      operationId: analyzeApplication
      tags: [analysis]
      summary: Analyze CV against job description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeApplicationBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"

  /applications/{id}/cover-letter:
    post:
      operationId: generateCoverLetter
      tags: [analysis]
      summary: Generate an ATS-friendly cover letter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCoverLetterBody"
      responses:
        "200":
          description: Cover letter content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterResult"

  /parse-cv:
    post:
      operationId: parseCv
      tags: [analysis]
      summary: Parse raw CV text into structured JSON using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseCvBody"
      responses:
        "200":
          description: Parsed CV structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseCvResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: AI parsing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /upload-cv:
    post:
      operationId: uploadCv
      tags: [analysis]
      summary: Upload a CV file (PDF or DOCX), extract text, and parse structure with AI
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Extracted text and parsed CV structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadCvResult"
        "400":
          description: Missing file, invalid type, or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "413":
          description: File exceeds 10 MB size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Text extraction failed or file has no readable content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - error

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id

    AuthUserEnvelope:
      type: object
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
      required:
        - user

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
          format: uri
        state:
          type: string
        nonce:
          type: string

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ParsedWorkExperience:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        start_date:
          type: string
        end_date:
          type: string
          nullable: true
        bullets:
          type: array
          items:
            type: string
      required:
        - company
        - title
        - start_date
        - end_date
        - bullets

    ParsedEducation:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
          nullable: true
        start_date:
          type: string
          nullable: true
        end_date:
          type: string
          nullable: true
      required:
        - institution
        - degree

    SaveTailoredCvBody:
      type: object
      properties:
        tailoredCvText:
          type: string
          minLength: 1
      required:
        - tailoredCvText

    SaveCoverLetterBody:
      type: object
      properties:
        coverLetterText:
          type: string
          minLength: 1
      required:
        - coverLetterText

    ParsedJobDescription:
      type: object
      properties:
        required_skills:
          type: array
          items:
            type: string
        preferred_skills:
          type: array
          items:
            type: string
        required_experience_years:
          type: number
          nullable: true
        key_responsibilities:
          type: array
          items:
            type: string
        must_have:
          type: array
          items:
            type: string
        nice_to_have:
          type: array
          items:
            type: string
        job_type:
          type: string
          enum: [full-time, part-time, contract, internship]
          nullable: true
        location_type:
          type: string
          enum: [remote, hybrid, onsite]
          nullable: true
      required:
        - required_skills
        - preferred_skills
        - key_responsibilities
        - must_have
        - nice_to_have

    ParsedCv:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        work_experience:
          type: array
          items:
            $ref: "#/components/schemas/ParsedWorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/ParsedEducation"
        skills:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
      required:
        - name
        - email
        - phone
        - location
        - summary
        - work_experience
        - education
        - skills
        - certifications
        - languages

    Application:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        jobTitle:
          type: string
        company:
          type: string
        jobDescription:
          type: string
        originalCvText:
          type: string
        parsedCvJson:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ParsedCv"
            - type: 'null'
        parsedJdJson:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ParsedJobDescription"
            - type: 'null'
        tailoredCvText:
          type: string
          nullable: true
        coverLetterText:
          type: string
          nullable: true
        keywordMatchScore:
          type: number
          nullable: true
        missingKeywords:
          type: array
          items:
            type: string
        matchedKeywords:
          type: array
          items:
            type: string
        missingInfoQuestions:
          type: array
          items:
            type: string
        sectionSuggestions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, analyzed, exported]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - jobTitle
        - company
        - jobDescription
        - originalCvText
        - missingKeywords
        - matchedKeywords
        - missingInfoQuestions
        - sectionSuggestions
        - status
        - createdAt
        - updatedAt

    CreateApplicationBody:
      type: object
      properties:
        userId:
          type: string
        jobTitle:
          type: string
        company:
          type: string
        jobDescription:
          type: string
        originalCvText:
          type: string
        parsedCvJson:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ParsedCv"
            - type: 'null'
      required:
        - userId
        - jobTitle
        - company
        - jobDescription
        - originalCvText

    UpdateApplicationBody:
      type: object
      properties:
        jobTitle:
          type: string
        company:
          type: string
        jobDescription:
          type: string
        originalCvText:
          type: string
        parsedCvJson:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ParsedCv"
            - type: 'null'
        tailoredCvText:
          type: string
        coverLetterText:
          type: string
        keywordMatchScore:
          type: number
        missingKeywords:
          type: array
          items:
            type: string
        matchedKeywords:
          type: array
          items:
            type: string
        missingInfoQuestions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, analyzed, exported]

    AnalyzeApplicationBody:
      type: object
      properties:
        confirmedAnswers:
          type: object
          additionalProperties:
            type: string

    AnalysisResult:
      type: object
      properties:
        tailoredCvText:
          type: string
        keywordMatchScore:
          type: number
        missingKeywords:
          type: array
          items:
            type: string
        matchedKeywords:
          type: array
          items:
            type: string
        missingInfoQuestions:
          type: array
          items:
            type: string
        sectionSuggestions:
          type: array
          items:
            type: string
      required:
        - tailoredCvText
        - keywordMatchScore
        - missingKeywords
        - matchedKeywords
        - missingInfoQuestions
        - sectionSuggestions

    GenerateCoverLetterBody:
      type: object
      properties:
        tone:
          type: string
          enum: [professional, enthusiastic, concise]
          default: professional
        additionalContext:
          type: string

    CoverLetterResult:
      type: object
      properties:
        coverLetterText:
          type: string
      required:
        - coverLetterText

    ParseCvBody:
      type: object
      properties:
        rawText:
          type: string
          minLength: 50
          maxLength: 100000
      required:
        - rawText

    ParseCvResult:
      type: object
      properties:
        parsedCv:
          $ref: "#/components/schemas/ParsedCv"
        parsedText:
          type: string
        sections:
          type: array
          items:
            type: string
      required:
        - parsedCv
        - parsedText
        - sections

    UploadCvResult:
      type: object
      properties:
        extractedText:
          type: string
        parsedCv:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ParsedCv"
            - type: 'null'
        fileName:
          type: string
        fileSize:
          type: number
      required:
        - extractedText
        - fileName
        - fileSize