# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ParsePilot AI API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ApplicationStatus } from "./applicationStatus";
import type { ParsedCv } from "./parsedCv";
import type { ParsedJobDescription } from "./parsedJobDescription";

export interface Application {
  id: string;
  userId: string;
  jobTitle: string;
  company: string;
  jobDescription: string;
  originalCvText: string;
  parsedCvJson?: ParsedCv | null;
  parsedJdJson?: ParsedJobDescription | null;
  tailoredCvText?: string | null;
  coverLetterText?: string | null;
  keywordMatchScore?: number | null;
  missingKeywords: string[];
  matchedKeywords: string[];
  missingInfoQuestions: string[];
  sectionSuggestions: string[];
  status: ApplicationStatus;
  createdAt: Date;
  updatedAt: Date;
}