# lib/api-zod/src/generated/types/parsedJobDescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ParsePilot AI API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ParsedJobDescriptionJobType } from "./parsedJobDescriptionJobType";
import type { ParsedJobDescriptionLocationType } from "./parsedJobDescriptionLocationType";

export interface ParsedJobDescription {
  required_skills: string[];
  preferred_skills: string[];
  required_experience_years?: number | null;
  key_responsibilities: string[];
  must_have: string[];
  nice_to_have: string[];
  job_type?: ParsedJobDescriptionJobType;
  location_type?: ParsedJobDescriptionLocationType;
}