# lib/api-zod/src/generated/types/saveTailoredCvBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ParsePilot AI API specification
 * OpenAPI spec version: 0.2.0
 */

export interface SaveTailoredCvBody {
  /** @minLength 1 */
  tailoredCvText: string;
}